backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langgraph>=0.0.15
langchain-community>=0.0.20
langchain-anthropic>=0.0.5
pandas>=2.2.0
openai>=1.12.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0 